+++ test/fixtures/cookbooks/test/recipes/system_install.rb
@@ -1,4 +1,4 @@
-# Install pyenv to the system path e.g. /usr/local/rbenv
+# Install pyenv to the system path e.g. /usr/local/pyenv
 pyenv_system_install 'system' do
   update_pyenv false
 end