+++ test/fixtures/cookbooks/test/recipes/user_install.rb
@@ -1,0 +1,13 @@
+# Install pyenv and makes it avilable to the selected user
+# version = '2.7.12'
+
+# Keeps the pyenv install upto date
+pyenv_user_install 'vagrant'
+
+# pyenv_python version do
+#   user 'vagrant'
+# end
+
+# pyenv_global version do
+#   user 'vagrant'
+# end