+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.0.0 (BREAKING CHANGES!!)
 Refactor and update the legacy code base
 * update `system_install` to be a resource
+* update `user_install` to be a resource
 * update integration test style
 
 ## 0.2.0